Commit 13
dbConnection.php:
Reformated code formation, created a working instance of data drawn from my database using the user table. Thinking the rests of the project through I will be restarting the tables, and creating them in a different way that creates a less complicated menu system.

menu.php
Removed the 3 placeholder collumns and added in a connection to the database statement that creates collumns from data in the table.

SQL files:
Small changes and beginning of a new table script for a more efficient system.