Commit 19
dbConnection.php:
Minimal changes made. removal of one line that would indicate a connection to the database has been made.

dbMenu.php:
A new file that contains the function for the menu to open on the page following the table entry.

menu.php:
Subtracted entire menu function and put into it's own file. Created a section for the basket of items to go.

Todo:
Return button, this goes back to the table number entry.

Basket buttons, to add and remove items from the basket.