Commit 26.3
Admin Menu Loader:
Loads the admin version of the menu with buttons that will soon have functionality to: edit, remove and add to or from the menu.

Integrated database openings to just one file with a couple to-be-used functions that are still being worked on.

Deleted previous db file connections.

Late-added class files for menu items and order items.

Refactors to file names and small objects, as well as code line changes for easier reading. Formatted code for easier reading.