Commit 27.3
Removed most of the SQL statements IN the code. Replaced with call methods to the database procedures.  Created a join statement to view in depth orders. Half the work towards placing an order, and it showing ALL data, in the database is done - difficult to solve the second half.

Next to do is for customers to find their orders.